Split session lifecycle tests
