+++ scripts/file-length-baseline.json
@@ -20,7 +20,6 @@
   "src/backend/services/session/service/chat/chat-message-handlers/handlers/load-session.handler.test.ts": 1598,
   "src/backend/services/session/service/lifecycle/session.config.service.test.ts": 1159,
   "src/backend/services/session/service/lifecycle/session.config.service.ts": 1080,
-  "src/backend/services/session/service/lifecycle/session.lifecycle.service.test.ts": 1620,
   "src/backend/services/session/service/lifecycle/session.lifecycle.service.ts": 1699,
   "src/backend/services/session/service/lifecycle/session.service.test.ts": 2777,
   "src/backend/services/workspace/resources/workspace.accessor.ts": 1055,
